refactor(storage): reworked the warning capture code
Refs #13
